% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{logo}
\alias{logo}
\title{Add ARPC Logo to ggplot}
\usage{
logo(position = "bottom-right", size = 0.1, path = NULL, alpha = 0.8)
}
\arguments{
\item{position}{Character, position of logo. Options: "bottom-right", "bottom-left",
"top-right", "top-left", "bottom", "top", "left", "right" (default: "bottom-right")}

\item{size}{Numeric, size of logo as proportion of plot area (default: 0.1)}

\item{path}{Character, path to logo file. If NULL, uses package default logo (default: NULL)}

\item{alpha}{Numeric, transparency of logo (0 = transparent, 1 = opaque) (default: 0.8)}
}
\value{
A ggplot2 annotation layer that can be added with +
}
\description{
Adds the ARPC logo to a ggplot object. This function returns a ggplot2 layer
that can be added to any ggplot using the + operator.
}
\examples{
library(ggplot2)

# Add logo to bottom-right (default)
ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point() +
  theme_arpc() +
  logo()
  
# Add logo to top-left corner
ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point() +
  theme_arpc() +
  logo(position = "top-left")
  
# Customize size and transparency
ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point() +
  theme_arpc() +
  logo(position = "top-right", size = 0.15, alpha = 0.6)
}
