% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{ndsu_colors}
\alias{ndsu_colors}
\title{NDSU Brand Colors}
\format{
A named character vector of hex color codes
}
\usage{
ndsu_colors(names = NULL)
}
\description{
A collection of colors from the official NDSU brand palette.
These colors are automatically applied when using theme_arpc().
}
\examples{
ndsu_colors()
ndsu_colors("green")
ndsu_colors(c("green", "yellow"))
}
